adding requested changes
