+++ src/qilisdk/analog/linear_schedule.py
@@ -23,57 +23,6 @@ from qilisdk.yaml import yaml
 class LinearSchedule(Schedule):
     """A Schedule that linearly interpolates between defined time steps."""
 
-    def get_coefficient(self, time_step: float, hamiltonian_key: str) -> Number:
-        t = time_step / self.dt
-        t_idx = int(t)
-
-        # if exactly defined, return directly
-        if t_idx in self._schedule and hamiltonian_key in self._schedule[t_idx]:
-            val = self._schedule[t_idx][hamiltonian_key]
-            return (
-                val.evaluate({}) if isinstance(val, Term) else (val.evaluate() if isinstance(val, Parameter) else val)
-            )
-
-        # search backwards for last defined
-        prev_idx, prev_val = None, None
-        for i in range(t_idx, -1, -1):
-            if i in self._schedule and hamiltonian_key in self._schedule[i]:
-                prev_idx = i
-                val = self._schedule[i][hamiltonian_key]
-                prev_val = (
-                    val.evaluate({})
-                    if isinstance(val, Term)
-                    else (val.evaluate() if isinstance(val, Parameter) else val)
-                )
-                break
-
-        # search forwards for next defined
-        next_idx, next_val = None, None
-        for i in range(t_idx + 1, int(self.T / self.dt) + 1):
-            if i in self._schedule and hamiltonian_key in self._schedule[i]:
-                next_idx = i
-                val = self._schedule[i][hamiltonian_key]
-                next_val = (
-                    val.evaluate({})
-                    if isinstance(val, Term)
-                    else (val.evaluate() if isinstance(val, Parameter) else val)
-                )
-                break
-
-        # cases
-        if prev_val is None and next_val is None:
-            return 0
-        if prev_val is None and next_val is not None:
-            return next_val
-        if next_val is None and prev_val is not None:
-            return prev_val
-
-        # linear interpolation
-        if next_idx is None or prev_idx is None or prev_val is None or next_val is None:
-            raise ValueError("Something unexpected happened while retrieving the coefficient.")
-        alpha: float = (t - prev_idx) / (next_idx - prev_idx)
-        return (1 - alpha) * prev_val + alpha * next_val
-
     def get_coefficient_expression(self, time_step: float, hamiltonian_key: str) -> Number | Term:
         t = time_step / self.dt
         t_idx = int(t)
@@ -110,6 +59,10 @@ class LinearSchedule(Schedule):
             raise ValueError("Something unexpected happened while retrieving the coefficient.")
         alpha: float = (t - prev_idx) / (next_idx - prev_idx)
         return (1 - alpha) * prev_expr + alpha * next_expr
+
+    def get_coefficient(self, time_step: float, hamiltonian_key: str) -> Number:
+        val = self.get_coefficient_expression(time_step=time_step, hamiltonian_key=hamiltonian_key)
+        return val.evaluate({}) if isinstance(val, Term) else (val.evaluate() if isinstance(val, Parameter) else val)
 
     def __getitem__(self, time_step: int) -> Hamiltonian:
         ham = Hamiltonian()
